% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export.tsHydro}
\alias{export.tsHydro}
\title{Export output}
\usage{
export.tsHydro(x, filename = "ts.dat", exportPar = FALSE)
}
\arguments{
\item{x}{An object of class "tsHydro"}

\item{filename}{Name of output file}

\item{exportPar}{Logic variable to specify if the estimated model parameters are saved to a file "tsPar.dat".}
}
\value{
The following elements
\itemize{
\item{oupfile("ts.dat") } A text file that contains three colunms;
"time", "wl", "wlsd". "time" is the time of each pass, where the water
level is estimated. "wl" is the estimated water level and "wlsd" is
the standard deviation of the estimated water level.
\item{"tspar.dat" } A text file that contains the optimized model parameters
}
}
\description{
This function saves the predicted water levels to a file
}
\examples{
\dontrun{
data(lakelevels)
fit <- get.TS(lakelevels)
export.tsHydro(fit, file = "myTS.dat", exportPar = TRUE)
}
}
\keyword{plot}
