% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lakelevels}
\alias{lakelevels}
\title{Lake-level satellite altimetry data}
\format{
A data frame with the following columns:
\describe{
\item{height}{Observed water level (m).}
\item{lake}{Lake identifier (integer).}
\item{track}{Satellite track identifier (integer).}
\item{time}{Observation time in decimal years.}
}
}
\usage{
lakelevels
}
\description{
Satellite altimetry observations of a small lake, suitable for testing
\code{\link[=get.TS]{get.TS()}}. 5,433 observations.
}
\keyword{datasets}
