% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.tsHydro.R
\name{plot.tsHydro}
\alias{plot.tsHydro}
\title{A Plot function}
\usage{
\method{plot}{tsHydro}(
  x,
  addRawDat = TRUE,
  addLine = TRUE,
  addError = FALSE,
  zoomOut = FALSE,
  lwd = 4,
  col = "blue",
  ...
)
}
\arguments{
\item{x}{Object returned by get.TS()}

\item{addRawDat}{Adds the data, which the estimated water levels are based on}

\item{addLine}{To add a line between the points that represents the estimated water levels.}

\item{addError}{To add error bars}

\item{zoomOut}{To zoom out. This option creates a plot which displays the range of the water level data.}

\item{lwd}{Line width}

\item{col}{Color}

\item{...}{Additional argumants to plot}
}
\description{
Plot an object returned by the function get.TS()
}
\examples{
\dontrun{
data(lakelevels)
fit <- get.TS(lakelevels)
plot(fit, addError = TRUE, col = 'blue')
plot(fit, zoomOut = TRUE, col = 'red')
}
}
\keyword{plot}
