% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.tsHydro.R
\name{print.tsHydro}
\alias{print.tsHydro}
\title{Simple print  of output}
\usage{
\method{print}{tsHydro}(x, ...)
}
\arguments{
\item{x}{An object of class "tsHydro"}

\item{...}{Unused, for S3 consistency with \code{\link[=print]{print()}}.}
}
\value{
Print the objective function and state convergence
}
\description{
This function presents a summary of the output
}
