% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.tsHydro.R
\name{summary.tsHydro}
\alias{summary.tsHydro}
\title{Summary of output}
\usage{
\method{summary}{tsHydro}(object, ...)
}
\arguments{
\item{object}{An object of class "tsHydro"}

\item{...}{Unused, for S3 consistency with \code{\link[=summary]{summary()}}.}
}
\value{
Summary of output
}
\description{
This function presents a summary of the output
}
