% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsHydroV2-package.R
\docType{package}
\name{tsHydroV2-package}
\alias{tsHydroV2}
\alias{tsHydroV2-package}
\title{tsHydroV2: Reconstruct water level time series from satellite altimetry data}
\description{
Reconstruct water level time series from satellite altimetry data using a state-space model implemented with RTMB.
}
\author{
\strong{Maintainer}: Karina Nielsen \email{karni@space.dtu.dk}

}
\keyword{internal}
